+++ src/ids-locale/cultures/cs-CZ.js
@@ -1,0 +1,64 @@
+// Info Comes from http://www.unicode.org/Public/cldr/25/
+const locale = {
+  // layout/language
+  language: 'cs',
+  englishName: 'Czech (Czech Republic)',
+  nativeName: 'čeština (Česká republika)',
+  // layout/orientation/@characters
+  direction: 'left-to-right',
+  // ca-gregorian
+  calendars: [{
+    name: 'gregorian',
+    // ca-gregorian/main/dates/calendars/gregorian/dateFormats/
+    dateFormat: {
+      separator: '.', // Infered
+      timeSeparator: ':',
+      short: 'dd.MM.yyyy', // use four digit year
+      medium: 'd. M. yyyy',
+      long: 'd. MMMM yyyy',
+      full: 'EEEE d. MMMM yyyy',
+      month: 'd. MMMM',
+      year: 'MMMM yyyy',
+      dayOfWeek: 'EEE d',
+      timestamp: 'H:mm:ss',
+      timestampMillis: 'H:mm:ss.sss',
+      hour: 'H:mm',
+      datetime: 'dd.MM.yyyy H:mm',
+      datetimeMillis: 'dd.MM.yyyy H:mm:ss.sss',
+      timezone: 'dd.MM.yyyy H:mm zz',
+      timezoneLong: 'dd.MM.yyyy H:mm zzzz'
+    }, // Infered short + short gregorian/dateTimeFormats
+    // ca-gregorian/main/dates/calendars/gregorian/days/format/short or abbreviated (2 digit)
+    days: {
+      wide: ['neděle', 'pondělí', 'úterý', 'středa', 'čtvrtek', 'pátek', 'sobota'],
+      abbreviated: ['ne', 'po', 'út', 'st', 'čt', 'pá', 'so'],
+      narrow: ['N', 'P', 'Ú', 'S', 'Č', 'P', 'S']
+    },
+    // ca-gregorian/main/dates/calendars/gregorian/months/format/wide
+    months: {
+      wide: ['leden', 'únor', 'březen', 'duben', 'květen', 'červen', 'červenec', 'srpen', 'září', 'říjen', 'listopad', 'prosinec'],
+      abbreviated: ['led', 'úno', 'bře', 'dub', 'kvě', 'čvn', 'čvc', 'srp', 'zář', 'říj', 'lis', 'pro']
+    },
+    // ca-gregorian/main/dates/calendars/gregorian/timeFormats/short
+    timeFormat: 'H:mm',
+    // ca-gregorian/main/dates/calendars/gregorian/dayPeriods/wide
+    dayPeriods: ['AM', 'PM'],
+    firstDayofWeek: 1 // Starts on Mon
+  }],
+  // numbers/currencyFormats-numberSystem-latn/standard
+  currencySign: 'Kč',
+  currencyFormat: '### ¤',
+  // numbers/symbols-numberSystem-latn
+  numbers: {
+    percentSign: '%',
+    percentFormat: '### %',
+    percentSuffix: ' %',
+    percentPrefix: '',
+    minusSign: '-',
+    decimal: ',',
+    group: ' ',
+    groupSizes: [3, 3]
+  }
+};
+
+export { locale };