+++ src/ids-locale/ids-locale-global.js
@@ -1,0 +1,6 @@
+import IdsLocale from './ids-locale';
+
+// Stores the global Locale instance.
+const locale = new IdsLocale();
+
+export default locale;