+++ src/ids-popup/ids-popup.js
@@ -163,8 +163,8 @@ class IdsPopup extends mix(IdsElement).with(
   }
 
   /**
-   * Return the properties we handle as getters/setters
-   * @returns {Array} The properties in an array
+   * Return the attributes we handle as getters/setters
+   * @returns {Array} The attributes in an array
    */
   static get attributes() {
     return [...super.attributes, ...POPUP_PROPERTIES];