+++ src/ids-theme-switcher/ids-theme-switcher.js
@@ -82,8 +82,8 @@ class IdsThemeSwitcher extends mix(IdsElement).with(IdsEventsMixin) {
   }
 
   /**
-   * Return the properties we handle as getters/setters
-   * @returns {Array} The properties in an array
+   * Return the attributes we handle as getters/setters
+   * @returns {Array} The attributes in an array
    */
   static get attributes() {
     return [...super.attributes, attributes.MODE, attributes.VERSION];