+++ src/ids-trigger-field/ids-trigger-field.js
@@ -49,8 +49,8 @@ class IdsTriggerField extends mix(IdsElement).with(IdsEventsMixin, IdsThemeMixin
   }
 
   /**
-   * Return the properties we handle as getters/setters
-   * @returns {Array} The properties in an array
+   * Return the attributes we handle as getters/setters
+   * @returns {Array} The attributes in an array
    */
   static get attributes() {
     return [attributes.TABBABLE, attributes.APPEARANCE, attributes.DISABLE_EVENTS];