+++ doc/CHANGELOG.md
@@ -64,6 +64,10 @@
   - Triggering element for the modal is now configurable with the `target` prop
   - Events are now just plain JS events
   - Markup has been changed to custom element `<ids-modal></ids-modal>`
+- `[NotificationBanner]` The Notification Banner is now a web component called ids-notification-banner
+  - The type, alert icon, message text, link and link text are all configurable via properties on the custom element
+  - Additionally, user's can define the notification via a JS api, where the `parent` and `id` can be configured
+  - Custom element: `<ids-notification-banner></ids-notification-banner>`
 - `[Pager]` Pager is now a web component called ids-pager ([WC#148](https://github.com/infor-design/enterprise-wc/issues/148))
   - there is a new base-component: `<ids-pager>`, and several sub-components to provide specific controls.
   - content is automatically managed/aligned via nesting controls inside of `<ids-pager-section>`
@@ -85,6 +89,8 @@
   - Markup has changed to a custom element `<ids-progress-bar></ids-progress-bar>`
   - If using events, events are now plain JS events.
   - Can now be imported as a single JS file and used with encapsulated styles
+- `[Locale]` The locale component has been changed to a mixin and added to many but not all components. You can now change the locale on the `ids-container` element to have it propagate down.
+  - See the [locale components docs](../src/ids-locales/README.md) for more detailed change info the the "Converting from Previous Versions" sections
 - `[Radio]` The Radio component has been changed to a web component and use with ids-radio and ids-radio-group.
   - Markup has changed to a custom element `<ids-radio-group><ids-radio></ids-radio></ids-radio-group>`
   - If using events, events are now plain JS events.